# perf(store): Batch save `Block`