Refs https://github.com/cometbft/cometbft/issues/1040

LSMs don't handle big batches very well ([1](https://github.com/cosmos/iavl/issues/619)). The ideal for `goleveldb` is ~ 100KB ([2](https://github.com/ethereum/go-ethereum/pull/15115)), which is more or less equal to our block part size (64kB).

Here are the results from running the Anton's benchmark on my local PC:

<details>

<summary>Benchmark</summary>

```
func BenchmarkBlockStore_SaveBlockWithExtendedCommit(b *testing.B) {
	dbTypes := []dbm.BackendType{
		dbm.GoLevelDBBackend,
	}
	NTXs := int64(1024)

	for _, dbType := range dbTypes {
		state, bs, cleanup := makeStateAndBlockStoreSpecificBackend(dbType)
		b.Run(fmt.Sprintf("DBType:%s;NTxs:%d", dbType, NTXs),
			func(b *testing.B) {
				for i := 0; i < b.N; i++ {
					height := bs.Height() + 1
					block := state.MakeBlock(height, test.MakeNTxs(height, NTXs), new(types.Commit), nil, state.Validators.GetProposer().Address)
					validPartSet, err := block.MakePartSet(types.BlockPartSizeBytes)
					if err != nil {
						b.Error(err)
					}
					seenCommit := makeTestExtCommit(block.Header.Height, cmttime.Now())

					bs.SaveBlockWithExtendedCommit(block, validPartSet, seenCommit)
				}
			})

		cleanup()
	}

}

type cleanupFunc func()

func makeStateAndBlockStoreSpecificBackend(backendType dbm.BackendType) (sm.State, *BlockStore, cleanupFunc) {
	config := test.ResetTestRoot("blockchain_reactor_test")
	blockDB, err := dbm.NewDB("block", backendType, config.DBDir())
	if err != nil {
		panic(fmt.Errorf("error creating %s: %w", backendType, err))
	}
	stateDB, err := dbm.NewDB("state", backendType, config.DBDir())
	if err != nil {
		panic(fmt.Errorf("error creating %s: %w", backendType, err))
	}
	stateStore := sm.NewStore(stateDB, sm.StoreOptions{
		DiscardABCIResponses: false,
	})
	state, err := stateStore.LoadFromDBOrGenesisFile(config.GenesisFile())
	if err != nil {
		panic(fmt.Errorf("error constructing state from genesis file: %w", err))
	}
	return state, NewBlockStore(blockDB), func() { os.RemoveAll(config.RootDir) }
}
```

</details>

```
## batching / 1 MB block / goleveldb
155921071 ns/op
## NO batching / 1 MB block / goleveldb
134169606 ns/op
## batching / 10 MB block / goleveldb
1135896143 ns/op
## NO batching / 10 MB block / goleveldb
1096632817 ns/op
```

If we want to support blocks bigger than 100kB on average, we should not batch single blocks. This becomes more important for big blocks (>1MB). The current max size is 4MB. This implies that CometBFT must handle such a scenario efficiently, which won't be the case w/ a single batch.

Now, the current block size of Cosmos Hub is 64kB, with peaks ranging from 200kB to 700kB. For Osmosis, it's 74kB. If we want to target the current customers, then creating a single batch makes sense.

The solution presented here is to batch save the whole block only if it's within 640kB (10 block parts * 64kB each). Otherwise, save each block part individually. All the commits are still saved as part of the batch.

---

#### PR checklist

- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use [unclog](https://github.com/informalsystems/unclog) to manage our changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code comments