@lasarojc 2023-12-07
Some coments not really related to this PR and a non-required suggestion to change the code. Otherwise, LGTM